+++ projects/cds/src/profiletag/cms-components/index.ts
@@ -1,0 +1,1 @@
+export * from './profile-tag.component';
