+++ projects/cds/src/profiletag/model/index.ts
@@ -1,0 +1,1 @@
+export * from './profile-tag.model';
