+++ projects/cds/src/profiletag/services/index.ts
@@ -1,0 +1,1 @@
+export * from './profile-tag.injector';
